Extract utilities from download-media.ts to helpers
- Extract DiscordRateLimiter class to rateLimiter.ts (145 lines)
- Move media utils to mediaHelper.ts: normalizeDiscordUrl, isSpoiler,
  isAnimated, getStickerExtension, getValidatedExtension, CONTENT_TYPE_TO_EXT,
  VALID_URL_EXTENSIONS
- Move detectActualFileType, getFileTypeDirAsync to fileHelper.ts
- Replace non-null assertions (!) with fallbacks (|| 'unknown')

Final consolidation results:
- download-media.ts: 1946 → 1545 lines (-401 lines, -20.6%)
- mediaHelper.ts: 203 → 324 lines (+121 lines)
- fileHelper.ts: 285 → 362 lines (+77 lines)
- rateLimiter.ts: 145 lines (new)
